[DF-4712] fix UIScreenTransition; creating sprite from texture region
Former-commit-id: 7050c575c1b1b7a038ccaa3c29449c40a32caa2b

+++ Sources/Internal/Render/2D/TextBlockGraphicsRender.cpp
@@ -49,8 +49,8 @@ void TextBlockGraphicsRender::Prepare(Texture *texture /*= NULL*/)
 	TextBlockRender::Prepare(texture);
 	
 	isPredrawed = false;
-    Texture * fbo = Texture::CreateFBO(textBlock->cacheDx, (float32)textBlock->cacheDy, FORMAT_RGBA8888, Texture::DEPTH_NONE);
-    sprite = Sprite::CreateFromTexture(fbo, 0, 0, fbo->GetWidth(), fbo->GetHeight());
+    Texture * fbo = Texture::CreateFBO(textBlock->cacheDx, textBlock->cacheDy, FORMAT_RGBA8888, Texture::DEPTH_NONE);
+    sprite = Sprite::CreateFromTexture(fbo, 0, 0, (float32)fbo->GetWidth(), (float32)fbo->GetHeight());
 	if (sprite && sprite->GetTexture())
 	{
 		if (!textBlock->isMultilineEnabled)
